+++ sdk/tables/azure-data-tables/azure/data/tables/__init__.py
@@ -5,6 +5,7 @@
 # --------------------------------------------------------------------------
 from ._entity import TableEntity, EntityProperty, EdmType, EntityMetadata
 from ._error import RequestTooLargeError, TableTransactionError, TableErrorCode
+from ._generated.models import UserDelegationKey
 from ._table_shared_access_signature import generate_table_sas, generate_account_sas
 from ._table_client import TableClient
 from ._table_service_client import TableServiceClient
@@ -50,4 +51,5 @@ __all__ = [
     "TransactionOperation",
     "RequestTooLargeError",
     "EntityMetadata",
+    "UserDelegationKey",
 ]